-Normalized the method names on the Global. -Colony now should have all the information available. Delayed the serialization and writing of the structures data. -Changed the type of the character identifier to the correct type of Integer.

+++ NeoCom.Model/src/main/java-datamngmt/org/dimensinfin/eveonline/neocom/manager/AbstractManager.java
@@ -63,7 +63,7 @@ public abstract class AbstractManager implements ICollaboration, IJsonAngular {
 		return new ArrayList<ICollaboration>();
 	}
 
-	public long getCredentialIdentifier () {
+	public int getCredentialIdentifier () {
 		if ( null == credential )
 			throw new RuntimeException("RT [AbstractManager]> Credential is not set on current Manager. Bad initialization.");
 		return credential.getAccountId();
